Improve local search by using the number of token matched
Improve search by using 2 level of matching.
First the query tokens are matched to every title. If all the tokens are
present, then compute jaccard distance.

This fixes a nasty bug like (say):
We have a query **burning in the skies**. But our cache doesn't have
that song by Linkin Park. However there is a song named **Burning
Stars** which is not what we had queried. So, using only `Jaccard
Distance` is not good here as it gives a wrong result.

This is fixed by using 2 levels of matching as said earlier.

We try to see if all the tokens of the query match to certain title.
So, here the tokens `[burning, in, the, skies]` are tried to match fully.
But we don't have any song where all the tokens match. So, this gives an
empty result from the cache and we go and fetch the song from the internet.

Assuming the song is in the cache with title **Linkin Park - Burning In
The Skies (Official)**. We have the query **burning in the skies**. So,
here, all the query tokens match and we return the result.

Adding to **Scenario 2**, say we have another **live** version of the
song. So, scenario 2 won't do good, since it will return only the first
song where tokens match, which might be the live version of the song
itself.

To mitigate that,  we compute `Jaccard Distance` too and try to get the nearest
match among all the tokens that were matched.