Change cache directory to `playx/songs`
Change caching of songs to `~/.playx/song` because I want other
directories in `~.playx` for various activities like `~/.playx/logs` for
logs and `~/.playx/playxlist` for playlist.

**Note**:
This update might render your playx to not recognize songs already
present in `~/.playx/` because now it is seeing the songs from `songs`
directory by default. So, please move the files to that directory.