Add a very simple autoplaylist
Add `playx/playlist/autoplaylist.py` to create a baseline for
auto-playlist. The barebone code uses simple counter of each songs from
the logs and generates a playlist with **.playx** extension to be able
to used later.
In future, this will be replaced by more general Markov Chains that
make use of state machines and transitions of one songs to other.