Fix title refactoring function in stringutils
There was a very naive, yet nasty bug in string utils where the title of
the song is checked if it ends with `mp3`. Previously, the title was
with with compared the string `.mp3`.

However, the `.`(dot) part of the title gets removed while passing through the punctuation fixer
gets removed while passing through the punctuation fixer. So, it should
be `title.endswith('mp3')` instead of `title.endswith(.mp3)`.